refactor: 모달 로직 커스텀훅으로 분리
